Isolate file types
