+++ src/models/file/index.ts
@@ -1,0 +1,4 @@
+export * from './file-metadata';
+export * from './is-file-metadata';
+export * from './raw-file-metadata';
+export * from './to-file-metadata';
